refactor(parser): split the AST in category
